:recycle: 커뮤니티 게시글 command in 패키지 추가 (#284)
