+++ application/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/CommunityPostController.kt
@@ -1,7 +1,8 @@
 package backend.team.ahachul_backend.api.community.adapter.web.`in`
 
 import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.*
-import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post.*
+import backend.team.ahachul_backend.api.community.application.command.`in`.DeleteCommunityPostCommand
+import backend.team.ahachul_backend.api.community.application.command.`in`.GetCommunityPostCommand
 import backend.team.ahachul_backend.api.community.application.port.`in`.CommunityPostUseCase
 import backend.team.ahachul_backend.common.annotation.Authentication
 import backend.team.ahachul_backend.common.dto.PageInfoDto
