+++ application/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/SearchCommunityPostDto.kt
@@ -1,6 +1,6 @@
 package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto
 
-import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post.SearchCommunityPostCommand
+import backend.team.ahachul_backend.api.community.application.command.`in`.SearchCommunityPostCommand
 import backend.team.ahachul_backend.api.community.domain.model.CommunityCategoryType
 import backend.team.ahachul_backend.common.domain.model.RegionType
 import backend.team.ahachul_backend.common.domain.model.YNType
