+++ application/src/main/kotlin/backend/team/ahachul_backend/api/community/application/port/in/CommunityPostUseCase.kt
@@ -1,7 +1,7 @@
 package backend.team.ahachul_backend.api.community.application.port.`in`
 
 import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.*
-import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post.*
+import backend.team.ahachul_backend.api.community.application.command.`in`.*
 import backend.team.ahachul_backend.common.dto.PageInfoDto
 
 interface CommunityPostUseCase {
