--- core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/GetCommunityPostCommand.kt
@@ -1,6 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
-
-class GetCommunityPostCommand(
-    val id: Long,
-) {
-}
