+++ core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/in/CreateCommunityPostCommand.kt
@@ -1,15 +1,14 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
+package backend.team.ahachul_backend.api.community.application.command.`in`
 
 import backend.team.ahachul_backend.api.community.domain.model.CommunityCategoryType
 import org.springframework.web.multipart.MultipartFile
 
-class UpdateCommunityPostCommand(
-    val id: Long,
+class CreateCommunityPostCommand(
     val title: String,
     val content: String,
     val categoryType: CommunityCategoryType,
     val hashTags: List<String> = arrayListOf(),
-    val uploadFiles: List<MultipartFile> = arrayListOf(),
-    val removeFileIds: List<Long> = arrayListOf()
+    val subwayLineId: Long,
+    val imageFiles: List<MultipartFile> = arrayListOf(),
 ) {
 }
