+++ core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/in/DeleteCommunityPostCommand.kt
@@ -1,0 +1,6 @@
+package backend.team.ahachul_backend.api.community.application.command.`in`
+
+class DeleteCommunityPostCommand(
+    val id: Long,
+) {
+}
