+++ core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/in/GetCommunityPostCommand.kt
@@ -1,0 +1,6 @@
+package backend.team.ahachul_backend.api.community.application.command.`in`
+
+class GetCommunityPostCommand(
+    val id: Long,
+) {
+}
