+++ core/src/main/kotlin/backend/team/ahachul_backend/api/community/application/command/in/UpdateCommunityPostCommand.kt
@@ -1,14 +1,15 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
+package backend.team.ahachul_backend.api.community.application.command.`in`
 
 import backend.team.ahachul_backend.api.community.domain.model.CommunityCategoryType
 import org.springframework.web.multipart.MultipartFile
 
-class CreateCommunityPostCommand(
+class UpdateCommunityPostCommand(
+    val id: Long,
     val title: String,
     val content: String,
     val categoryType: CommunityCategoryType,
     val hashTags: List<String> = arrayListOf(),
-    val subwayLineId: Long,
-    val imageFiles: List<MultipartFile> = arrayListOf(),
+    val uploadFiles: List<MultipartFile> = arrayListOf(),
+    val removeFileIds: List<Long> = arrayListOf()
 ) {
 }
